+++ core/src/main/scala/unionmirror/auto/package.scala
@@ -4,5 +4,5 @@ import scala.deriving.Mirror
 
 export unionmirror.UnionMirror.synth
 
-transparent inline given mirrorSumOf: [T] => Mirror.SumOf[T] =
+transparent inline given mirrorSumOf[T]: Mirror.SumOf[T] =
   ${ unionmirror.internal.UnionMirrorImpl.derivedUnionMirrorOf[T] }